Extract logic from EuiComboBox to shrink file length. - Extract matching logic into matchingOptions local service. - Extract EuiComboBoxInput and EuiComboBoxOptionsList child components.
