+++ src/components/combo_box/_index.scss
@@ -1,4 +1,3 @@
 @import 'combo_box';
-@import 'combo_box_pill';
-@import 'combo_box_option';
-@import 'combo_box_title';
+@import 'combo_box_input/index';
+@import 'combo_box_options_list/index';
