+++ src/components/combo_box/combo_box_input/_combo_box_pill.scss
