+++ src/components/combo_box/combo_box_input/_index.scss
@@ -1,0 +1,1 @@
+@import 'combo_box_pill';
