+++ src/components/combo_box/combo_box_input/index.js
@@ -1,0 +1,3 @@
+export {
+  EuiComboBoxInput,
+} from './combo_box_input';
