+++ src/components/combo_box/combo_box_options_list/_combo_box_option.scss
