+++ src/components/combo_box/combo_box_options_list/_index.scss
@@ -1,0 +1,2 @@
+@import 'combo_box_option';
+@import 'combo_box_title';
