+++ src/components/combo_box/combo_box_options_list/index.js
@@ -1,0 +1,3 @@
+export {
+  EuiComboBoxOptionsList,
+} from './combo_box_options_list';
