--- src/components/combo_box/combo_box_pill.js
@@ -1,52 +1,0 @@
-import React, { Component } from 'react';
-import PropTypes from 'prop-types';
-import classNames from 'classnames';
-
-import {
-  EuiIcon,
-} from '../icon';
-
-export class EuiComboBoxPill extends Component {
-  static propTypes = {
-    option: PropTypes.object.isRequired,
-    children: PropTypes.node,
-    className: PropTypes.string,
-    onClose: PropTypes.func.isRequired,
-  }
-
-  onClickClose = () => {
-    const { onClose, option } = this.props;
-    onClose(option);
-  }
-
-  render() {
-    const {
-      children,
-      className,
-      option, // eslint-disable-line no-unused-vars
-      onClose, // eslint-disable-line no-unused-vars
-      ...rest
-    } = this.props;
-
-    const classes = classNames('euiComboBoxPill', className);
-
-    return (
-      <div
-        className={classes}
-        {...rest}
-      >
-        <div className="euiComboBoxPill__label">
-          {children}
-        </div>
-
-        <button
-          className="euiComboBoxPill__close"
-          onClick={this.onClickClose}
-          tabIndex="-1"
-        >
-          <EuiIcon type="cross" className="euiComboBoxPill__closeIcon" />
-        </button>
-      </div>
-    );
-  }
-}
